high activity control
The eifficacy of NE spikes are compared with the efficacy of coincident spikes whith other neurons firing within the coincidence window. The number of spikes are subsampled to the same. The significant of change is compared for 2 halves of the data: pairs with increased efficacy for NE spikes and pairs with decreased efficacy for NE spikes. 2 plots: 1 showing the trend of change and use color to show NE spikes or coincident spikes; 1 showing significance on individual bases by comparing the mean of efficacy calculated with the subsampled spikes, with multiple comparison correction (p*=p x n_total_pairs)